Add files via upload
These datasets are used for analyzing the "Impact of Brexit on International Student Enrolment in UK Universities." In 4 datasets, one for student enrolment from EU  and another for non-EU region. The other two datasets, in which one focusing on England-based universities and the other on universities in  Scotland, Wales, and Northern Ireland.